address comments pt2
