+++ azext_edge/edge/util/az_client.py
@@ -187,7 +187,7 @@ def get_authz_client(subscription_id: str, **kwargs) -> "AuthorizationManagement
 def get_keyvault_client(subscription_id: str, keyvault_name: str, **kwargs) -> "SecretClient":
     from azure.keyvault.secrets import SecretClient
 
-    # TODO: this only supports azure public cloud only
+    # TODO: this only supports azure public cloud for now
     client = SecretClient(
         credential=AZURE_CLI_CREDENTIAL,
         subscription_id=subscription_id,
