+++ azext_edge/tests/edge/orchestration/test_deletion_unit.py
@@ -11,6 +11,7 @@ import pytest
 
 from azext_edge.edge.providers.orchestration.deletion import IoTOperationsResource
 from azext_edge.edge.providers.orchestration.work import IOT_OPS_EXTENSION_TYPE
+from azext_edge.tests.helpers import generate_ops_resource
 
 from ...generators import generate_random_string
 
@@ -62,20 +63,6 @@ def spy_deletion_manager(mocker):
     }
 
 
-def _generate_ops_resource(segments: int = 1) -> IoTOperationsResource:
-    resource_id = ""
-    for _ in range(segments):
-        resource_id = f"{resource_id}/{generate_random_string()}"
-
-    resource = IoTOperationsResource(
-        resource_id=resource_id,
-        display_name=resource_id.split("/")[-1],
-        api_version=generate_random_string(),
-    )
-
-    return resource
-
-
 def _assemble_resource_map_mock(
     resource_map_mock: Mock,
     extensions: Optional[List[dict]],
@@ -91,11 +78,7 @@ def _assemble_resource_map_mock(
         IOT_OPS_EXTENSION_TYPE: {"id": "aio-ext-id"}
     }
     resource_map_mock().extensions.append(
-        IoTOperationsResource(
-            resource_id="aio-ext-id",
-            display_name="aio-extension",
-            api_version="aio-ext-api"
-        )
+        IoTOperationsResource(resource_id="aio-ext-id", display_name="aio-extension", api_version="aio-ext-api")
     )
 
 
@@ -113,11 +96,11 @@ def _assemble_resource_map_mock(
         },
         {
             "resources": [
-                _generate_ops_resource(4),
+                generate_ops_resource(4),
             ],
-            "resource sync rules": [_generate_ops_resource()],
-            "custom locations": [_generate_ops_resource()],
-            "extensions": [_generate_ops_resource()],
+            "resource sync rules": [generate_ops_resource()],
+            "custom locations": [generate_ops_resource()],
+            "extensions": [generate_ops_resource()],
             "meta": {
                 "expected_total": 4,
                 "resource_batches": 1,
@@ -125,14 +108,14 @@ def _assemble_resource_map_mock(
         },
         {
             "resources": [
-                _generate_ops_resource(4),
-                _generate_ops_resource(4),
-                _generate_ops_resource(3),
-                _generate_ops_resource(1),
+                generate_ops_resource(4),
+                generate_ops_resource(4),
+                generate_ops_resource(3),
+                generate_ops_resource(1),
             ],
             "resource sync rules": [],
-            "custom locations": [_generate_ops_resource()],
-            "extensions": [_generate_ops_resource(), _generate_ops_resource()],
+            "custom locations": [generate_ops_resource()],
+            "extensions": [generate_ops_resource(), generate_ops_resource()],
             "meta": {
                 "expected_total": 7,
                 "resource_batches": 3,
@@ -198,10 +181,10 @@ def test_batch_resources(
             },
         },
         {
-            "resources": [_generate_ops_resource(4), _generate_ops_resource(2)],
-            "resource sync rules": [_generate_ops_resource(), _generate_ops_resource()],
-            "custom locations": [_generate_ops_resource()],
-            "extensions": [_generate_ops_resource()],
+            "resources": [generate_ops_resource(4), generate_ops_resource(2)],
+            "resource sync rules": [generate_ops_resource(), generate_ops_resource()],
+            "custom locations": [generate_ops_resource()],
+            "extensions": [generate_ops_resource()],
             "meta": {
                 "expected_total": 5,
                 "resource_batches": 2,
@@ -210,10 +193,10 @@ def test_batch_resources(
         },
         # Currently no associated custom location means no non-extensions get deleted
         {
-            "resources": [_generate_ops_resource(4), _generate_ops_resource(2)],
-            "resource sync rules": [_generate_ops_resource()],
+            "resources": [generate_ops_resource(4), generate_ops_resource(2)],
+            "resource sync rules": [generate_ops_resource()],
             "custom locations": [],
-            "extensions": [_generate_ops_resource()],
+            "extensions": [generate_ops_resource()],
             "meta": {
                 "expected_total": 4,
                 "resource_batches": 2,
@@ -224,7 +207,7 @@ def test_batch_resources(
             "resources": [],
             "resource sync rules": [],
             "custom locations": [],
-            "extensions": [_generate_ops_resource()],
+            "extensions": [generate_ops_resource()],
             "meta": {
                 "expected_total": 1,
                 "resource_batches": 0,
